requirements.txt
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
PyNaCl==1.6.1
PySide6==6.10.1
PyYAML==6.0.3
Requests
scikit-learn==1.7.2
setuptools
SQLAlchemy==2.0.44
substrate_interface==1.7.11
torch==2.9.1
web3==7.14.0
munch>=4.0.0
async-substrate-interface>=1.5.6
asyncstdlib~=3.13.0
colorama~=0.4.6
fastapi~=0.110.1
msgpack-numpy-opentensor~=0.5.0
nest_asyncio==1.6.0
netaddr==1.3.0
packaging
python-statemachine~=2.1
retry==0.9.2
uvicorn
wheel
